requirements.txt
telethon
python-dotenv
pandas
openai
cohere
pinecone==4.0.0
tenacity
icecream